Added, Draw Ellipse mode: angle setting can now be changed using "Rotate Clockwise" and "Rotate Counterclockwise" actions. Added, Draw Line and Draw Curve modes: added "Auto-finish drawing" setting. When enabled, the modes will automatically finish drawing when currently drawn lines and already existing level geometry form a closed shape. Changed: sector-wise linedef flipping is now done using the new "Align Linedefs" action. "Flip Linedefs" action works the same as in DB2 again. Changed: when a map was already loaded, using "Open Map" action will use that map's directory as the starting directory. Changed: official IWADs can no longer be saved. Changed: disabled lump ranges/duplicate entries checks for official IWADs. Changed: wad type is now preserved when saving a map (previously all wads were saved as PWADs). Changed: moved Updater.exe launch much closer to the editor termination point to avoid any chance of it closing the editor before it properly closes itself. Updated ZDoom_DECORATE.cfg (A_Blast). Updated documentation.

+++ branches/GZDoomBuilder/Source/Plugins/BuilderModes/VisualModes/BaseVisualGeometrySector.cs
@@ -580,7 +580,7 @@ namespace CodeImp.DoomBuilder.BuilderModes
 		}
 		
 		// Processing
-		public virtual void OnProcess(float deltatime)
+		public virtual void OnProcess(long deltatime)
 		{
 			// If the texture was not loaded, but is loaded now, then re-setup geometry
 			if(setuponloadedtexture != 0)
